+++ gui/sdlimgui/win_dbgscr.go
@@ -529,7 +529,7 @@ func (win *winDbgScr) setScaleFromWindow(sz imgui.Vec2) {
 
 func (win *winDbgScr) getScaling(horiz bool) float32 {
 	if horiz {
-		return float32(pixelWidth * win.scr.aspectBias * win.scaling)
+		return pixelWidth * win.scr.aspectBias * win.scaling
 	}
 	return win.scaling
 }
